backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
googlemaps==4.10.0
idna==3.6
numpy==1.26.2
Pillow==10.1.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
